+++ SentryTestUtils/TestOptions.swift
@@ -2,19 +2,27 @@ import Foundation
 import Sentry
 
 public extension Options {
-    @available(*, deprecated)
-    func setIntegrations(_ integrations: [AnyClass]) {
-        self.integrations = integrations.map {
-            NSStringFromClass($0)
-        }
-    }
     
-    @available(*, deprecated)
     func removeAllIntegrations() {
-        self.integrations = []
+        enableAutoSessionTracking = false
+        enableWatchdogTerminationTracking = false
+        enableAutoPerformanceTracing = false
+        enableCrashHandler = false
+        swiftAsyncStacktraces = false
+        enableAppHangTracking = false
+        enableNetworkTracking = false
+        enableNetworkBreadcrumbs = false
+        enableCaptureFailedRequests = false
+        enableAutoBreadcrumbTracking = false
+        enableCoreDataTracing = false
+        enableFileIOTracing = false
+        #if (os(iOS) || os(tvOS) || (swift(>=5.9) && os(visionOS))) && !SENTRY_NO_UIKIT
+        enableUserInteractionTracing = false
+        attachViewHierarchy = false
+        enableUIViewControllerTracing = false
+        #endif
     }
-    
-    @available(*, deprecated)
+
     static func noIntegrations() -> Options {
         let options = Options()
         options.removeAllIntegrations()