+++ Sources/Sentry/SentryTimeToDisplayTracker.m
@@ -2,6 +2,7 @@
 
 #if SENTRY_HAS_UIKIT
 
+#    import "SentryInternalDefines.h"
 #    import "SentryLogC.h"
 #    import "SentryOptions+Private.h"
 #    import "SentryProfilingConditionals.h"
@@ -199,7 +200,12 @@
 
 - (void)addTimeToDisplayMeasurement:(id<SentrySpan>)span name:(NSString *)name
 {
-    NSTimeInterval duration = [span.timestamp timeIntervalSinceDate:span.startTimestamp] * 1000;
+    if (!span.startTimestamp) {
+        return;
+    }
+    NSTimeInterval duration =
+        [span.timestamp timeIntervalSinceDate:SENTRY_UNWRAP_NULLABLE(NSDate, span.startTimestamp)]
+        * 1000;
     [span setMeasurement:name value:@(duration) unit:SentryMeasurementUnitDuration.millisecond];
 }