+++ Tests/SentryTests/Integrations/Screenshot/SentryScreenshotIntegrationTests.swift
@@ -41,31 +41,28 @@ class SentryScreenshotIntegrationTests: XCTestCase {
         clearTestState()
     }
 
-    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func test_attachScreenshot_disabled() {
         SentrySDK.start {
+            $0.removeAllIntegrations()
             $0.attachScreenshot = false
-            $0.setIntegrations([SentryScreenshotIntegration.self])
         }
         XCTAssertEqual(SentrySDKInternal.currentHub().getClient()?.attachmentProcessors.count, 0)
         XCTAssertFalse(sentrycrash_hasSaveScreenshotCallback())
     }
     
-    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func test_attachScreenshot_enabled() {
         SentrySDK.start {
+            $0.removeAllIntegrations()
             $0.attachScreenshot = true
-            $0.setIntegrations([SentryScreenshotIntegration.self])
         }
         XCTAssertEqual(SentrySDKInternal.currentHub().getClient()?.attachmentProcessors.count, 1)
         XCTAssertTrue(sentrycrash_hasSaveScreenshotCallback())
     }
     
-    @available(*, deprecated, message: "This is deprecated because SentryOptions integrations is deprecated")
     func test_uninstall() {
         SentrySDK.start {
+            $0.removeAllIntegrations()
             $0.attachScreenshot = true
-            $0.setIntegrations([SentryScreenshotIntegration.self])
         }
         SentrySDK.close()